chore: move all folders from shared/ to lib/
