+++ lib/main.dart
@@ -3,7 +3,7 @@ import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:get_storage/get_storage.dart';
 
 import 'app.dart';
-import 'shared/dependencies.dart';
+import 'utils/dependencies.dart';
 
 void main() async {
   registerDependencies();
