+++ lib/mocks/mock_reviews.dart
@@ -1,4 +1,4 @@
-import '../extensions.dart';
+import '../utils/extensions.dart';
 import '../models/review.dart';
 import 'mock_users.dart';
