+++ lib/models/shop.dart
