+++ lib/providers/shop_search_name_provider.dart
